feat(islands): cameraFocusHelper.ts added to correct ortho zoom. Scaling added to ortho zoom and perspective framing to maintain consistency in camera usage across model sizes.
